ai-arena/backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
openai
google-generativeai
